refactor(editor-template-plugins): use shared title component
