+++ packages/editor/src/plugins/serlo-template-plugins/course/course-page.tsx
@@ -3,16 +3,10 @@ import {
   type EditorPluginProps,
   string,
 } from '@editor/plugin'
-import { focus, useAppDispatch } from '@editor/store'
-import { useEditorStrings } from '@serlo/frontend/src/contexts/logged-in-data-context'
-import { useEffect, useRef } from 'react'
+import { useEffect } from 'react'
 
-import {
-  entity,
-  editorContent,
-  entityType,
-  headerInputClasses,
-} from '../common/common'
+import { entity, editorContent, entityType } from '../common/common'
+import { EntityTitleInput } from '../common/entity-title-input'
 import { ToolbarMain } from '../toolbar-main/toolbar-main'
 
 export const coursePageTypeState = entityType(
@@ -40,40 +34,16 @@ function CoursePageTypeEditor(
   props: EditorPluginProps<CoursePageTypePluginState, { skipControls: boolean }>
 ) {
   const { title, content, icon } = props.state
-  const titleRef = useRef<HTMLInputElement>(null)
-
-  const dispatch = useAppDispatch()
 
   useEffect(() => {
     // setting not used any more, reset to explanation for now
     if (icon.value !== 'explanation') icon.set('explanation')
   })
 
-  useEffect(() => {
-    // focus on title, remove focus from content
-    setTimeout(() => {
-      dispatch(focus(null))
-      titleRef.current?.focus()
-    })
-    // only after creating plugin
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  const placeholder = useEditorStrings().templatePlugins.coursePage.title
-
   return (
     <>
       <article>
-        <h1 className="serlo-h1 mt-12">
-          <input
-            ref={titleRef}
-            autoFocus
-            className={headerInputClasses}
-            placeholder={placeholder}
-            value={title.value}
-            onChange={(e) => title.set(e.target.value)}
-          />
-        </h1>
+        <EntityTitleInput title={title} forceFocus />
 
         {content.render()}
